requirements.txt
paho-mqtt
oscpy
pyyaml